Move GraphPath and GraphPathIterable to a different package
